+++ openaev-api/src/main/java/io/openaev/api/threat_arsenal/ThreatArsenalApi.java
@@ -1,11 +1,14 @@
 package io.openaev.api.threat_arsenal;
 
 import static io.openaev.config.TenantUriUtils.TENANT_PREFIX;
+import static io.openaev.rest.settings.PreviewFeature.INJECT_CHAINING;
 
 import io.openaev.aop.AccessControl;
 import io.openaev.api.asset.dto.SecurityPlatformSimpleOutput;
 import io.openaev.api.threat_arsenal.dto.*;
 import io.openaev.database.model.Action;
+import io.openaev.database.model.ChainingTypeRegistry;
+import io.openaev.database.model.PrimitiveType;
 import io.openaev.database.model.ResourceType;
 import io.openaev.rest.injector_contract.InjectorContractService;
 import io.openaev.rest.injector_contract.input.InjectorContractSearchPaginationInput;
@@ -13,6 +16,7 @@ import io.openaev.rest.injector_contract.output.InjectorContractAuthorCountOutpu
 import io.openaev.rest.injector_contract.output.InjectorContractBaseOutput;
 import io.openaev.rest.injector_contract.output.InjectorContractDomainCountOutput;
 import io.openaev.schema.model.PropertySchemaDTO;
+import io.openaev.service.PreviewFeatureService;
 import io.openaev.service.threat_arsenal.ThreatArsenalService;
 import io.openaev.utils.mapper.SecurityPlatformMapper;
 import io.openaev.utils.pagination.SearchPaginationInput;
@@ -38,6 +42,7 @@ public class ThreatArsenalApi {
   public static final String TENANT_THREAT_ARSENAL_URL = TENANT_PREFIX + "/threat_arsenals";
 
   private final ThreatArsenalService threatArsenalService;
+  private final PreviewFeatureService previewFeatureService;
 
   // -- READ --
 
@@ -49,6 +54,19 @@ public class ThreatArsenalApi {
       resourceType = ResourceType.THREAT_ARSENAL)
   public ThreatArsenalActionFullOutput threatArsenal(@PathVariable String actionId) {
     return threatArsenalService.findById(actionId);
+  }
+
+  @Operation(
+      summary = "Get all primitive chaining types",
+      description = "Returns primitive types available for payload arguments.")
+  @GetMapping({
+    THREAT_ARSENAL_URL + "/argument-types/",
+    TENANT_THREAT_ARSENAL_URL + "/argument-types/"
+  })
+  @Transactional
+  @AccessControl(actionPerformed = Action.SEARCH, resourceType = ResourceType.THREAT_ARSENAL)
+  public List<PrimitiveType> getArgumentTypes() {
+    return resolveAvailableTypes();
   }
 
   @Operation(summary = "Get filterable property schemas for threat arsenal")
@@ -151,6 +169,13 @@ public class ThreatArsenalApi {
       @PathVariable String actionId) {
     return SecurityPlatformMapper.toSimpleOutputs(
         threatArsenalService.getSecurityPlatformsForActionRemediation(actionId));
+  }
+
+  private List<PrimitiveType> resolveAvailableTypes() {
+    if (!previewFeatureService.isFeatureEnabled(INJECT_CHAINING)) {
+      return List.of(PrimitiveType.Text, PrimitiveType.Document, PrimitiveType.TargetedAsset);
+    }
+    return ChainingTypeRegistry.getPrimitiveTypes();
   }
 
   // -- CREATE --
