+++ openaev-front/src/admin/components/threat_arsenal/form/useArgumentTypes.ts
@@ -1,6 +1,6 @@
 import { useEffect, useMemo, useState } from 'react';
 
-import { fetchArgumentTypes } from '../../../../actions/chaining/chaining-actions';
+import fetchArgumentTypes from '../../../../actions/payloads/payload-argument-actions';
 
 type UseArgumentTypesResult = {
   argumentTypes: string[];
